fix promocode

+++ lib/features/promocode/data/data_source/promo_code_data_source.dart
@@ -10,7 +10,7 @@ class PromoCodeDataSource {
 
   Future<CodeModel> addPromoCode({
     required String code,
-    required int discount,
+    required double discount,
     required String expireDate,
     required int maxUsage,
     required int used,
